+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.4.1] - 2023-10-23
+### Added
+- Hybrid spell casting saves.
+- Skill cards display mastery level.
+### Changed
+- Skill card formatting.
+
 ## [0.4.0] - 2023-09-20
 ### Added
 - Token base display.